Move resize retry test to its own file for file-length budget
The retry regression test pushed CLINotifyProcessIntegrationRegressionTests.swift
over the Swift file-length budget (workflow-guard-tests). Move it into a new
SSHPTYResizeRetryIntegrationTests.swift that extends the same test class, so it
still reuses the mock-socket / bundled-CLI harness helpers while keeping the
large file unchanged. Wired into the cmuxTests target.
